# api/openapi.yaml
openapi: 3.1.0
info:
  title: Catalift API
  version: 1.7.0
  description: |
    The REST API the Catalift web app uses: sellers upload products and start generation, reviewers edit,
    regenerate, approve and export listings, and both see AI cost and the budget state (ADR-0007).
    Serves: US-00-001 to US-00-012. Style: docs/api/README.md. The server default for X-API-Version is the
    newest date version, currently 2026-10-04.
  contact:
    name: Catalift team
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  x-api-style: >-
    REST over HTTPS with JSON bodies, plus multipart for CSV and image uploads and text/csv for export
    downloads; one Go service exposing five areas (auth, catalogue, listings, export, cost) under one /v1.
  x-versioning: >-
    Major in the path (/v1, ADR-0007); dated behaviour changes inside a major in X-API-Version, server
    default the newest (2026-10-04).
  x-authentication: >-
    A server-side session in an HTTP-only secure cookie (catalift_session) set by POST /v1/sessions, with
    an X-CSRF-Token header on every state-changing request (ADR-0006). Roles seller and reviewer are
    checked on every route; a seller calling a reviewer action gets 403 forbidden_role.
  x-conventions:
    - rule: 'The base path is /v1, not /api/v1.'
      why: 'ADR-0007 and HLD section 5 fixed /v1; the reverse proxy forwards /v1 and serves the web app at /.'
...
    - rule: 'Ids are decimal strings of the database''s bigint ids.'
...
servers:
...
security:
  - sessionCookie: []
...
paths:
...
  /brands:
    get:
      operationId: listBrands
      summary: List brands with their voice notes
      description: With filter[upload_id], only brands named by that upload's products (v1.7).
      tags: [brands]
      x-story-ids: [US-00-003]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/UploadFilter'
      responses:
        '200':
          description: One page of brands, by name
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BrandPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /brands/{brand_id}:
    parameters:
      - $ref: '#/components/parameters/BrandId'
    patch:
      operationId: updateBrand
      summary: Set a brand's voice (seller)
      description: >-
        Replaces the tone (voice_note) and, when given, the words to avoid; sending the same values twice leaves the
        same state. Words are trimmed, blanks and repeats (ignoring case) dropped; at most 50, each at most 60
        characters. A new list re-checks the brand's listings: a word to avoid in one is a brand_avoid_word failure,
        and an approved listing that now fails loses its approval (v1.6, ADR-0012).
      tags: [brands]
      x-story-ids: [US-00-003]
      x-idempotent: true
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BrandUpdate' }
            example: { voice_note: Warm and plain-spoken; no slang; mention fabric first., words_to_avoid: [cheap, best ever] }
      responses:
        '200':
          description: The updated brand
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Brand' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /uploads:
    post:
      operationId: createUpload
      summary: Upload a product CSV
      description: >-
        Multipart CSV with columns sku, category, brand, price. Creates one product per valid row and reports
        rejected rows with their reasons, including a SKU repeated in the file or already in Catalift (D21).
        Brands named for the first time are created.
      tags: [uploads]
      x-story-ids: [US-00-001]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/UploadCreate' }
      responses:
        '201':
          description: Upload summary
          headers:
            Location: { schema: { type: string, format: uri } }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Upload' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /uploads/latest:
    get:
      operationId: getLatestUpload
      summary: Get the newest upload's summary
      description: >-
        The screens show only the newest upload's records; older uploads stay in the database. 404 before any upload
        (v1.7).
      tags: [uploads]
      x-story-ids: [US-00-001]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: The newest upload's summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Upload' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /uploads/{upload_id}:
    parameters:
      - $ref: '#/components/parameters/UploadId'
    get:
      operationId: getUpload
      summary: Get an upload's summary
      tags: [uploads]
      x-story-ids: [US-00-001, US-00-011]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: The upload summary with its rejected rows and AI cost total
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Upload' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /products:
    get:
      operationId: listProducts
      summary: List products with detection status and AI cost
      tags: [products]
      x-story-ids: [US-00-001, US-00-002, US-00-011]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: filter[upload_id]
          in: query
          description: Only products from this upload (served by idx_products_upload_id).
          schema: { type: string, pattern: '^[0-9]+$' }
        - name: filter[detection_status]
          in: query
          description: Only products in this detection state (sequential scan by design, under 10^4 rows).
          schema: { $ref: '#/components/schemas/DetectionStatus' }
      responses:
        '200':
          description: One page of products, by SKU
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
  /products/{product_id}:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      operationId: getProduct
      summary: Get one product
      tags: [products]
      x-story-ids: [US-00-001, US-00-002, US-00-011]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: The product with its attributes, images and AI cost
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    patch:
      operationId: updateProduct
      summary: Correct a product's details (seller)
      description: >-
        Re-checks SKU, category, brand and price with the upload's rules and saves them, or says why not (v1.5). A
        product whose listings a reviewer approved can no longer change (422).
      tags: [products]
      x-story-ids: [US-00-001]
      x-idempotent: true
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RowFix' }
      responses:
        '200':
          description: Saved, or refused with the reason
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductSaveResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /products/{product_id}/enrich:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    post:
      operationId: enrichProduct
      summary: Run the vision call again for one product (seller)
      description: >-
        Resets the product's detection, queues every channel's listing again and makes one enrich-v1 vision call for
        this product only (v1.5). It counts toward the budget; refused while the budget is blocked or the product has
        no photo (422).
      tags: [products]
      x-story-ids: [US-00-002, US-00-003]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: The run that tracks the call
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerationRun' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /products/{product_id}/image:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      operationId: getProductImage
      summary: Get a product's thumbnail
      description: The product's first photo as its 1024 px JPEG detection copy, for the review grid (v1.2).
      tags: [products]
      x-story-ids: [US-00-007]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
      responses:
        '200':
          description: The thumbnail
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /products/{product_id}/images:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    post:
      operationId: createProductImages
      summary: Add photos to one product (seller)
      description: >-
        Attaches the photos to this product whatever the files are called: the fix for a product left without a photo
        (v1.4). The same type, size and pixel checks as an upload apply.
      tags: [products]
      x-story-ids: [US-00-001]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/ImagesCreate' }
      responses:
        '201':
          description: The photos attached and those refused
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImagesResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '500': { $ref: '#/components/responses/InternalError' }
  /row-errors:
    get:
      operationId: listRowErrors
      summary: Rejected CSV rows not yet fixed
      description: >-
        Every row an upload refused and the seller has not corrected yet, with the values as typed and the categories a
        correction may use (v1.4).
      tags: [uploads]
      x-story-ids: [US-00-001]
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - name: filter[upload_id]
          in: query
          schema: { type: string, pattern: '^[0-9]+$' }
      responses:
        '200':
          description: The open rejected rows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RowErrorList' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: discardRowErrors
      summary: Discard every rejected row (seller)
      description: >-
        Drops every open rejected row, or one upload's with filter[upload_id], so the seller can clear rows they will
        not load. Nothing else changes; discarding twice discards nothing the second time (v1.5).
      tags: [uploads]
      x-story-ids: [US-00-001]
      x-idempotent: true
      parameters:
        - $ref: '#/components/parameters/ApiVersion'
        - $ref: '#/components/parameters/CsrfToken'
        - name: filter[upload_id]
          in: query
          schema: { type: string, pattern: '^[0-9]+$' }
      responses:
        '200':
          description: How many rows were discarded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscardResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: catalift_session
      description: Server-side session set by POST /v1/sessions (ADR-0006).
  parameters:
    ApiVersion:
      name: X-API-Version
      in: header
      required: false
      description: Date version, e.g. 2026-10-04. Missing means the server default (the newest).
      schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: The csrf_token from the session; required on every state-changing request (ADR-0006).
      schema: { type: string }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from page.next_cursor.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    BrandId:
      name: brand_id
      in: path
      required: true
      schema: { type: string, pattern: '^[0-9]+$' }
...
    UploadFilter:
      name: filter[upload_id]
      in: query
      description: Only this upload's records; the screens pass the latest upload (v1.7).
      schema: { type: string, pattern: '^[0-9]+$' }
...
  headers:
    XRequestId:
      schema: { type: string }
      description: Echoed or generated request id; quote it in support tickets.
  schemas:
...
    BrandPage:
      type: object
      required: [data, page]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Brand' } }
        page: { $ref: '#/components/schemas/Page' }
      example:
        data:
          - { id: '3', name: Indigo Loom, voice_note: null, words_to_avoid: [], created_at: 2026-10-04T08:00:00Z, updated_at: 2026-10-04T08:00:00Z }
        page: { next_cursor: null, has_more: false }
...
  responses:
    BadRequest:
      description: Malformed request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: No session, or wrong email or password
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Wrong role or missing CSRF token
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: No such record
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    TooManyRequests:
      description: Rate limited; see Retry-After
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    InternalError:
      description: Unexpected failure; request_id is in the body
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }